+++ lib/src/public/main/collection/group_channel_message_collection/base_message_collection.dart
@@ -6,6 +6,7 @@ import 'package:collection/collection.dart';
 import 'package:sendbird_chat_sdk/sendbird_chat_sdk.dart';
 import 'package:sendbird_chat_sdk/src/internal/main/chat/chat.dart';
 import 'package:sendbird_chat_sdk/src/internal/main/chat_cache/cache_service.dart';
+import 'package:sendbird_chat_sdk/src/internal/main/chat_manager/collection_manager/auto_resend_manager.dart';
 import 'package:sendbird_chat_sdk/src/internal/main/chat_manager/collection_manager/collection_manager.dart';
 import 'package:sendbird_chat_sdk/src/internal/main/chat_manager/db_manager.dart';
 import 'package:sendbird_chat_sdk/src/internal/main/logger/sendbird_logger.dart';
@@ -397,6 +398,8 @@ abstract class BaseMessageCollection {
         throw exception;
       }
     }
+
+    AutoResendManager().startAutoResend(chat);
   }
 
   void _setValuesForInitialize({